Button functionality
This is the first javascript code I've added to the project, and it's
basically just query selectors to get reference to all the buttons and
event listeners that do nothing but change the text in the calculator
screen ('screen' div) depending on what button is pressed. Actual
calculation will be added further on.